Correctly identify deeply analyzed artifacts
The recursion was going the wrong way. Instead of following the dependencies
downwards (i.e. towards dependees), the recursive step should go upwards
(i.e. towards dependents). This way it can identify which other artifacts
can be finished due to the fact that one of its dependees is now finished.