Remove redundant completed artifacts and useless values from persistence
Instances of could reference other instances (as their dependees). Because of this the same artifact could occur repeatedly (once on its own and many more times as a dependee of other artifacts), which led to bloated YAML files.

The YAML output used to contain entries for null fields and empty collections. This made the file needlessly verbose and bloated so it was removed.