Defer writing of results (YAML and JWOS) to the end of the computation
Once the results become larger it is very wasteful to write them often. It would've been preferable to push all result writing to the end of the computation but it does make sense to log intermediate results (in cases of crashes). This intermediate result writing should maybe occur less often, though.